Splits files into new folder
Creates helper file for existing methods
